Extract base 64 to static
