+++ pom.xml
@@ -4,7 +4,7 @@
 
   <groupId>org.mindrot</groupId>
   <artifactId>jbcrypt</artifactId>
-  <version>0.4</version>
+  <version>0.5-SNAPSHOT</version>
   <packaging>jar</packaging>
 
   <name>jbcrypt</name>
